Cambio id entidades de Integer a Long Nuevos métodos para buscar inspecciones con autocompletar por num insp o nombre unidad  - Insp. no finalizadas sin solic. no finalizadas  - Insp. no finalizadas sin solic. no finalizadas y sin cuest. no finalizados Cambio método busqueda solicitudes por correo a no finalizadas (usado en provisionalSolicitud) Añadidos campos busquedaSolicitud: nº inspeccion, ambito, nombre unidad, tipo inspeccion Cambio flujo solicitudes nueva validacion Jefe Equipo antes del envío Nuevos campos entidad SolicitudDocumentacionPrevia, fechaNoConforme, usernameNoConforme Cambio flujo solicitudes nueva alternativa a finalizar solicitud, declarar no conforme  - reactiva userProv  - elimina fechaCumplimentacion  - graba fechaNoConforme

+++ src/main/java/es/mira/progesin/persistence/entities/gd/GestDocSolicitudDocumentacion.java
@@ -36,10 +36,10 @@ public class GestDocSolicitudDocumentacion implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE)
 	@Column(name = "id", nullable = false)
-	private Integer id;
+	private Long id;
 
 	@Column(name = "idSolicitud")
-	private Integer idSolicitud;
+	private Long idSolicitud;
 
 	@Column(name = "extension", nullable = false, length = 4)
 	private String extension;
